# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.CreateSoundRequest:
    properties:
      album:
        example: Summer Vibes
        type: string
      genre:
        example: Electronic
        type: string
      name:
        example: My Awesome Sound
        type: string
    type: object
  handlers.DeleteSoundRequest:
    properties:
      name:
        example: Sound to delete
        type: string
    type: object
...
  handlers.UpdateSoundRequest:
    properties:
      album:
        example: Updated Album
        type: string
      genre:
        example: Updated Genre
        type: string
      name:
        example: Updated Sound Name
        type: string
    type: object
  handlers.UploadRequest:
    properties:
      filename:
        example: mysound.mp3
        type: string
      full_path:
        example: /app/static/uploads/mysound.mp3
        type: string
      message:
        example: file uploaded successfully
        type: string
      path:
        example: uploads/mysound.mp3
        type: string
    type: object
  sound.SoundDTO:
    properties:
      album:
        type: string
      author_id:
        type: integer
      duration:
        type: integer
      file_format:
        type: string
      file_name:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      genre:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: string
      upload_date:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Audio sharing and streaming platform API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SoundTube API
  version: "1.0"
paths:
...
  /api/sounds:
    get:
      description: Get a list of all available sounds
      produces:
      - application/json
      responses:
        "200":
          description: List of sounds
          schema:
            items:
              $ref: '#/definitions/sound.SoundDTO'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all sounds
      tags:
      - sounds
    post:
      consumes:
      - application/json
      description: Create a new sound record (without file upload)
      parameters:
      - description: Sound data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateSoundRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Sound created successfully
          schema:
            type: string
        "400":
          description: Invalid input format
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create sound
      tags:
      - sounds
  /api/sounds/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a sound by name
      parameters:
      - description: Sound ID
        in: path
        name: id
        required: true
        type: integer
      - description: Delete data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.DeleteSoundRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Sound deleted successfully
          schema:
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - not sound owner
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Sound not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete sound
      tags:
      - sounds
    patch:
      consumes:
      - application/json
      description: Update sound information
      parameters:
      - description: Sound ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update sound data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateSoundRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Sound updated successfully
          schema:
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - not sound owner
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Sound not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update sound
      tags:
      - sounds
...
  /api/sounds/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload an audio file for an existing sound record
      parameters:
      - description: Audio file to upload
        in: formData
        name: file
        required: true
        type: file
      - description: Sound name to associate with file
        in: formData
        name: name
        required: true
        type: string
      - description: Upload sound file
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UploadRequest'
      produces:
      - application/json
      responses:
        "400":
          description: Missing file or sound name
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: File upload or database update failed
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload sound file
      tags:
      - sounds
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"